refactor(delegation): split vc-verifier into types + stateless checks + class
Extract the result/resolver types into vc-verifier.types.ts (re-exported) and the basic/signature/status checks into vc-verification-checks.ts as pure functions; DelegationCredentialVerifier now holds only orchestration + its result cache. Class name/public API unchanged. 517 -> 226/245/100, all <300; 1607 tests green.
