+++ src/delegation/vc-verifier.ts
@@ -8,116 +8,31 @@
  * Stage 2: Parallel advanced checks (signature, status)
  * Stage 3: Combined results
  *
+ * The stateless checks live in `./vc-verification-checks.ts`; the shapes in
+ * `./vc-verifier.types.ts` (re-exported below). This class owns orchestration
+ * and the per-instance result cache.
+ *
  * Related Spec: KYA-OS §4.3, W3C VC Data Model 1.1
  */
 
+import type { DelegationCredential } from "../types/protocol.js";
 import type {
-  DelegationCredential,
-  CredentialStatus,
-} from "../types/protocol.js";
+  DelegationVCVerificationResult,
+  VerifyDelegationVCOptions,
+  DIDResolver,
+  StatusListResolver,
+  StatusCheckResult,
+  SignatureVerificationFunction,
+} from "./vc-verifier.types.js";
 import {
-  isDelegationCredentialExpired,
-  isDelegationCredentialNotYetValid,
-  validateDelegationCredential,
-} from "../types/protocol.js";
+  validateBasicProperties,
+  verifySignature,
+  checkCredentialStatus,
+} from "./vc-verification-checks.js";
 
-/**
- * Properties a DelegationCredential `credentialSubject` may carry. Anything
- * else is treated as a claim-intended field and rejected by default — a
- * permission credential must not smuggle claims (KYA-OS §11.6).
- */
-const DELEGATION_SUBJECT_KEYS: readonly string[] = ["id", "delegation"];
-
-/**
- * Machine-readable outcome of the credential-status check. Both values still
- * DENY (fail-closed) — the label only tells the consumer *why*, so retry vs
- * re-consent can be decided:
- * - `revoked`: the status bit is set — a settled deny. Re-consent required.
- * - `status_unresolvable`: the status could not be read (resolver missing,
- *   threw, or infra unavailable). Deny now, but the credential may still be
- *   good; safe to retry once status resolution recovers.
- */
-export type StatusOutcome = "revoked" | "status_unresolvable";
-
-export interface DelegationVCVerificationResult {
-  valid: boolean;
-  reason?: string;
-  /**
-   * Present only when a status check produced a deny. Distinguishes a settled
-   * `revoked` from a transient `status_unresolvable` without weakening
-   * fail-closed — both still deny (see {@link StatusOutcome}).
-   */
-  statusOutcome?: StatusOutcome;
-  stage: "basic" | "signature" | "status" | "complete";
-  cached?: boolean;
-  metrics?: {
-    basicCheckMs?: number;
-    signatureCheckMs?: number;
-    statusCheckMs?: number;
-    totalMs: number;
-  };
-  checks?: {
-    basicValid?: boolean;
-    signatureValid?: boolean;
-    statusValid?: boolean;
-  };
-}
-
-export interface VerifyDelegationVCOptions {
-  skipCache?: boolean;
-  skipSignature?: boolean;
-  skipStatus?: boolean;
-  didResolver?: DIDResolver;
-  statusListResolver?: StatusListResolver;
-}
-
-export interface DIDResolver {
-  resolve(did: string): Promise<DIDDocument | null>;
-}
-
-export interface DIDDocument {
-  '@context'?: string | string[];
-  id: string;
-  alsoKnownAs?: string[];
-  verificationMethod?: VerificationMethod[];
-  authentication?: (string | VerificationMethod)[];
-  assertionMethod?: (string | VerificationMethod)[];
-}
-
-export interface VerificationMethod {
-  id: string;
-  type: string;
-  controller: string;
-  publicKeyJwk?: unknown;
-  publicKeyBase58?: string;
-  publicKeyMultibase?: string;
-}
-
-export interface StatusListResolver {
-  checkStatus(status: CredentialStatus): Promise<boolean>;
-}
-
-/**
- * Internal result of a credential-status check. `outcome` is set on every
- * deny (see {@link StatusOutcome}); it is surfaced as
- * `DelegationVCVerificationResult.statusOutcome`.
- */
-interface StatusCheckResult {
-  valid: boolean;
-  reason?: string;
-  outcome?: StatusOutcome;
-  durationMs?: number;
-}
-
-export interface SignatureVerificationFunction {
-  (
-    vc: DelegationCredential,
-    publicKeyJwk: unknown,
-  ): Promise<{
-    valid: boolean;
-    reason?: string;
-  }>;
-}
+// Re-export the shared type surface so existing imports of DIDResolver /
+// SignatureVerificationFunction / StatusListResolver / … are unchanged.
+export * from "./vc-verifier.types.js";
 
 export class DelegationCredentialVerifier {
   private didResolver?: DIDResolver;
@@ -176,7 +91,7 @@ export class DelegationCredentialVerifier {
     }
 
     const basicCheckStart = Date.now();
-    const basicValidation = this.validateBasicProperties(vc);
+    const basicValidation = validateBasicProperties(vc);
     const basicCheckMs = Date.now() - basicCheckStart;
 
     if (!basicValidation.valid) {
@@ -196,7 +111,11 @@ export class DelegationCredentialVerifier {
     }
 
     const signaturePromise = !options.skipSignature
-      ? this.verifySignature(vc, options.didResolver || this.didResolver)
+      ? verifySignature(
+          vc,
+          options.didResolver || this.didResolver,
+          this.signatureVerifier,
+        )
       : Promise.resolve<{
           valid: boolean;
           reason?: string;
@@ -208,7 +127,7 @@ export class DelegationCredentialVerifier {
 
     const statusPromise =
       !options.skipStatus && vc.credentialStatus
-        ? this.checkCredentialStatus(
+        ? checkCredentialStatus(
             vc.credentialStatus,
             options.statusListResolver || this.statusListResolver,
           )
@@ -253,216 +172,6 @@ export class DelegationCredentialVerifier {
     }
 
     return result;
-  }
-
-  private validateBasicProperties(vc: DelegationCredential): {
-    valid: boolean;
-    reason?: string;
-  } {
-    const schemaValidation = validateDelegationCredential(vc);
-    if (!schemaValidation.success) {
-      return {
-        valid: false,
-        reason: `Schema validation failed: ${schemaValidation.error?.message}`,
-      };
-    }
-
-    if (isDelegationCredentialExpired(vc)) {
-      return { valid: false, reason: "Delegation credential expired" };
-    }
-
-    if (isDelegationCredentialNotYetValid(vc)) {
-      return { valid: false, reason: "Delegation credential not yet valid" };
-    }
-
-    const delegation = vc.credentialSubject.delegation;
-    if (delegation.status === "revoked") {
-      return { valid: false, reason: "Delegation status is revoked" };
-    }
-    if (delegation.status === "expired") {
-      return { valid: false, reason: "Delegation status is expired" };
-    }
-
-    if (!delegation.issuerDid || !delegation.subjectDid) {
-      return { valid: false, reason: "Missing issuer or subject DID" };
-    }
-
-    if (!vc.proof) {
-      return { valid: false, reason: "Missing proof" };
-    }
-
-    const subjectShape = this.validateSubjectShape(vc);
-    if (!subjectShape.valid) {
-      return subjectShape;
-    }
-
-    return { valid: true };
-  }
-
-  /**
-   * Reject a delegation credential whose `credentialSubject` carries properties
-   * other than `id` and `delegation`. Claim-bearing fields in a permission
-   * credential separate designation from authorization — the confused-deputy
-   * class (KYA-OS §11.6). Unconditional as of 1.4.0.
-   */
-  private validateSubjectShape(vc: DelegationCredential): {
-    valid: boolean;
-    reason?: string;
-  } {
-    const extraneous = Object.keys(vc.credentialSubject).filter(
-      (key) => !DELEGATION_SUBJECT_KEYS.includes(key),
-    );
-
-    if (extraneous.length === 0) {
-      return { valid: true };
-    }
-
-    return {
-      valid: false,
-      reason:
-        `credentialSubject contains non-delegation field(s): ${extraneous.join(", ")}. ` +
-        `A DelegationCredential subject MUST carry only 'id' and 'delegation' (KYA-OS §11.6).`,
-    };
-  }
-
-  private async verifySignature(
-    vc: DelegationCredential,
-    didResolver?: DIDResolver,
-  ): Promise<{ valid: boolean; reason?: string; durationMs?: number }> {
-    const startTime = Date.now();
-
-    try {
-      const issuerDid =
-        typeof vc.issuer === "string" ? vc.issuer : vc.issuer.id;
-
-      if (!didResolver || !this.signatureVerifier) {
-        return {
-          valid: false,
-          reason:
-            "No DID resolver or signature verifier configured — signature cannot be verified",
-          durationMs: Date.now() - startTime,
-        };
-      }
-
-      const didDoc = await didResolver.resolve(issuerDid);
-      if (!didDoc) {
-        return {
-          valid: false,
-          reason: `Could not resolve issuer DID: ${issuerDid}`,
-          durationMs: Date.now() - startTime,
-        };
-      }
-
-      if (!vc.proof) {
-        return {
-          valid: false,
-          reason: "Proof is missing",
-          durationMs: Date.now() - startTime,
-        };
-      }
-
-      const verificationMethodId = vc.proof["verificationMethod"];
-      if (!verificationMethodId) {
-        return {
-          valid: false,
-          reason: "Proof missing verificationMethod",
-          durationMs: Date.now() - startTime,
-        };
-      }
-
-      const verificationMethod = this.findVerificationMethod(
-        didDoc,
-        verificationMethodId as string,
-      );
-      if (!verificationMethod) {
-        return {
-          valid: false,
-          reason: `Verification method ${verificationMethodId} not found`,
-          durationMs: Date.now() - startTime,
-        };
-      }
-
-      const publicKeyJwk = verificationMethod.publicKeyJwk;
-      if (!publicKeyJwk) {
-        return {
-          valid: false,
-          reason: "Verification method missing publicKeyJwk",
-          durationMs: Date.now() - startTime,
-        };
-      }
-
-      const verificationResult = await this.signatureVerifier(vc, publicKeyJwk);
-
-      return {
-        valid: verificationResult.valid,
-        reason: verificationResult.reason,
-        durationMs: Date.now() - startTime,
-      };
-    } catch (error) {
-      return {
-        valid: false,
-        reason: `Signature verification error: ${error instanceof Error ? error.message : "Unknown error"}`,
-        durationMs: Date.now() - startTime,
-      };
-    }
-  }
-
-  /**
-   * Resolve a credential's revocation status, fail-closed. Every deny is
-   * labelled with a {@link StatusOutcome} so consumers can tell a settled
-   * `revoked` (re-consent) from a transient `status_unresolvable` (retry):
-   * a missing resolver or a resolver that throws both DENY as
-   * `status_unresolvable`; a set status bit DENIES as `revoked`.
-   */
-  private async checkCredentialStatus(
-    status: CredentialStatus,
-    statusListResolver?: StatusListResolver,
-  ): Promise<StatusCheckResult> {
-    const startTime = Date.now();
-
-    try {
-      if (!statusListResolver) {
-        return {
-          valid: false,
-          reason:
-            "Credential has credentialStatus but no status list resolver is configured — cannot verify revocation status",
-          outcome: "status_unresolvable",
-          durationMs: Date.now() - startTime,
-        };
-      }
-
-      const isRevoked = await statusListResolver.checkStatus(status);
-
-      if (isRevoked) {
-        return {
-          valid: false,
-          reason: `Credential revoked via StatusList2021 (${status.statusPurpose})`,
-          outcome: "revoked",
-          durationMs: Date.now() - startTime,
-        };
-      }
-
-      return {
-        valid: true,
-        durationMs: Date.now() - startTime,
-      };
-    } catch (error) {
-      return {
-        valid: false,
-        reason: `Status check error: ${error instanceof Error ? error.message : "Unknown error"}`,
-        outcome: "status_unresolvable",
-        durationMs: Date.now() - startTime,
-      };
-    }
-  }
-
-  private findVerificationMethod(
-    didDoc: DIDDocument,
-    verificationMethodId: string,
-  ): VerificationMethod | undefined {
-    return didDoc.verificationMethod?.find(
-      (vm) => vm.id === verificationMethodId,
-    );
   }
 
   private getFromCache(id: string): DelegationVCVerificationResult | null {
